Refactored vector stores to maybe support langchain
